Game now has a stored history of what plays were made and it is possible to scroll through this.
Game mode is now selectable, but the match feature is currently
unfinished. Implementing a match-n feature. Where the game will take
each card and compare rank and suit with the rest of the cards. The
match method will return the highest score from this.
Next feature will be adding new back and front images. And if time
permits, changing this to a more iOS7 look.